+++ sound-mind-studio/tests/main.cpp
@@ -13,6 +13,7 @@
 #include <QStandardPaths>
 #include <QtTest/QtTest>
 
+#include "test_audio_snippet_picker_dialog.h"
 #include "test_canvas_widget.h"
 #include "test_create_project_wizard.h"
 #include "test_landing_page.h"
@@ -70,6 +71,9 @@ int main(int argc, char** argv) {
 
     PlaybackPanelTest playbackPanelTest;
     status |= QTest::qExec(&playbackPanelTest, argc, argv);
+
+    AudioSnippetPickerDialogTest audioSnippetPickerDialogTest;
+    status |= QTest::qExec(&audioSnippetPickerDialogTest, argc, argv);
 
     return status;
 }