+++ sound-mind-studio/tests/test_audio_snippet_picker_dialog.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <QObject>
+
+class AudioSnippetPickerDialogTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void listsOneRowPerSnippetWithItsTimespan();
+    void allRowsStartChecked();
+    void selectedIndicesReturnsOnlyCheckedRows();
+    void selectAllCheckboxTogglesEveryRow();
+};